don't show close button on standalone image gallery
