--- common/views/components/ImageGallery/ImageGallery.config.js
@@ -1,30 +1,0 @@
-export const name = 'Image gallery';
-export const status = 'graduated';
-export const context = {
-  items: [
-    {
-      contentUrl: `https://wellcomecollection.files.wordpress.com/2017/06/to-01-08-l0062946.jpg`,
-      caption: `A Chinese woman suffering from leprosy. In the 1880s, the president of the Louisiana State Board of Health blamed the “filthy, vicious, debased… Chinese” for the disease.`,
-      width: 2000,
-      height: 3038,
-    },
-    {
-      contentUrl: `https://wellcomecollection.files.wordpress.com/2017/06/to-01-09-l0007201.jpg`,
-      caption: `Syphilis has been blamed on a range of different nationalities and religions, including the French.`,
-      width: 1088,
-      height: 1732,
-    },
-    {
-      contentUrl: `https://wellcomecollection.files.wordpress.com/2017/06/to-01-10-l0000611.jpg`,
-      caption: `Poster warning of the “alarming approach” of what was described as “Indian” cholera, produced in London in 1831.`,
-      width: 1198,
-      height: 1548,
-    },
-    {
-      contentUrl: `https://wellcomecollection.files.wordpress.com/2017/06/to-01-11-l0072276.jpg`,
-      caption: `Tuberculosis is often described as a disease of “the poor”. This 1951 poster by the Swiss Association Against Tuberculosis appealed for funds to “Help us to help the poor”.`,
-      width: 2000,
-      height: 2828,
-    },
-  ],
-};
